fix: sonar cloud duplications
