+++ exam-craft-extension-frontend/src/screens/codeGeneration/GenerationTestAtributesScreen.tsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react"
-import logoExamCraft from "../../../assets/icon512.png"
+// IMPORTANTE: Asegúrate de que la ruta sea la correcta donde guardaste tu nuevo componente
+import { Header } from "~src/components/Header" 
 import { parseMasterPrompt } from "~src/utils/promptParser"
 import { sendToGemini } from "~src/services/geminiService"
 import testAttributesPromptMarkdown from "bundle-text:../../prompts/generation-exam-repository/generation_tests.md"
@@ -98,7 +99,7 @@ ${restricciones}
     };
 
     const handleSaveToChrome = () => {
-        if (typeof chrome !== "undefined" && chrome.storage && chrome.storage.local) {
+        if (typeof chrome !== "undefined" && chrome.storage?.local) {
             
             if (!initialData?.project?.id) {
                 alert("Error: No se ha encontrado el ID del proyecto para actualizar.");
@@ -139,18 +140,6 @@ ${restricciones}
         URL.revokeObjectURL(url);
     };
 
-    const breadcrumbButtonStyle: React.CSSProperties = {
-        background: 'none',
-        border: 'none',
-        padding: 0,
-        margin: 0,
-        font: 'inherit',
-        color: '#4a3728',
-        cursor: 'pointer',
-        display: 'inline',
-        outline: 'none'
-    };
-
     const breadcrumbItems = [
         { label: 'INICIO', action: onWelcome },
         { label: 'CREAR EXAMEN', action: onCreateExam },
@@ -163,31 +152,12 @@ ${restricciones}
 
     return (
         <div className="exam-app">
-            <header className="app-header">
-                <div className="header-left">
-                    <button 
-                        type="button"
-                        className="logo-icon" 
-                        onClick={onWelcome} 
-                        style={{ background: 'none', border: 'none', padding: 0, cursor: 'pointer', outline: 'none' }}
-                        aria-label="Ir a inicio"
-                    >
-                        <img src={logoExamCraft} alt="Logo ExamCraft" width="60" height="60" />
-                    </button>
-
-                    <nav className="breadcrumb-nav">
-                        {breadcrumbItems.map((item) => (
-                            <React.Fragment key={item.label}>
-                                <button type="button" style={breadcrumbButtonStyle} onClick={item.action}>
-                                    {item.label}
-                                </button>
-                                <span className="breadcrumb-separator">{' > '}</span>
-                            </React.Fragment>
-                        ))}
-                        <span className="breadcrumb-current">TEST DE ATRIBUTOS</span>
-                    </nav>
-                </div>
-            </header>
+            {/* AQUÍ USAMOS NUESTRO NUEVO COMPONENTE COMPARTIDO */}
+            <Header 
+                onWelcome={onWelcome} 
+                breadcrumbItems={breadcrumbItems} 
+                currentStep="TEST DE ATRIBUTOS" 
+            />
 
             <main className="main-content"> 
                 <div className="wf-layout-container">
